requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.2.0=pypi_0
alsa-lib=1.2.3.2=h166bdaf_0
aom=3.3.0=h27087fc_1
attrs=22.1.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py310ha9d4c09_0
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.07.19=h06a4308_0
cairo=1.16.0=ha12eb4b_1010
certifi=2022.9.14=py310h06a4308_0
cffi=1.15.1=py310h74dc2b5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
contourpy=1.0.5=pypi_0
cryptography=37.0.2=py310h597c629_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pypi_0
dbus=1.13.6=h5008d03_3
expat=2.4.8=h27087fc_0
ffmpeg=4.4.1=hd7ab26d_2
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h8e229c2_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.37.3=pypi_0
freeglut=3.2.2=h9c3ff4c_1
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h27087fc_1009
giflib=5.2.1=h36c2ea0_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.18.5=hf529b03_3
gstreamer=1.18.5=h9f60fe5_3
harfbuzz=4.2.0=h40b6f09_0
hdf5=1.12.1=nompi_h2386368_104
icu=69.1=h9c3ff4c_0
idna=3.4=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=2.0.33=ha77e612_0
jbig=2.1=h7f98852_2003
jpeg=9e=h166bdaf_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=13.0.1=default_hc23dcda_0
libcurl=7.83.1=h7bff187_0
libdeflate=1.7=h7f98852_5
libdrm=2.4.113=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libglib=2.70.2=h174f98d_4
libglu=9.0.0=he1b5a44_1001
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm13=13.0.1=hf817b99_2
libnghttp2=1.47.0=h727a467_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopencv=4.5.5=py310h1897127_9
libopus=1.3.1=h7f98852_1
libpciaccess=0.16=h516909a_0
libpng=1.6.37=h21135ba_2
libpq=14.5=hd77ab85_0
libprotobuf=3.20.1=h6239696_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_1
libuuid=2.32.1=h7f98852_1000
libva=2.15.0=h166bdaf_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h885dcf4_1
libzlib=1.2.12=h166bdaf_3
llvm-openmp=14.0.4=he0ac6c6_0
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.6.0=pypi_0
mediapipe=0.8.11=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310ha2c4b55_0
mkl_fft=1.3.1=py310h2b4bcf5_1
mkl_random=1.2.2=py310h00e6091_0
mysql-common=8.0.27=ha770c72_3
mysql-libs=8.0.27=hfa10184_3
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
nspr=4.32=h9c3ff4c_1
nss=3.77=h2350873_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.23.1=py310h1794996_0
numpy-base=1.23.1=py310hcba007f_0
opencv=4.5.5=py310hff52083_9
opencv-contrib-python=4.6.0.66=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py310h6a678d5_0
pcre=8.45=h9c3ff4c_0
pillow=9.2.0=py310hace64e9_1
pip=22.1.2=py310h06a4308_0
pixman=0.40.0=h36c2ea0_0
protobuf=3.20.2=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=4.5.5=py310hfdc917e_9
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py310h06a4308_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.5=h582c2e5_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310h06a4308_0
qt=5.12.9=ha98a1a1_5
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_1
setuptools=63.4.1=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.2=h5082296_0
svt-av1=0.9.1=h27087fc_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py310_cu116
torchmetrics=0.10.0=pyhd8ed1ab_0
torchvision=0.13.1=py310_cu116
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022c=h04d1e81_0
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h5eee18b_3
zstd=1.5.0=ha95c52a_0
